Add some function key mappings
The one set of mapping deals with typing in graves in insert mode.
Since ` is used as my tmux prefix, I used to always have to mash it
twice to send it through. Now I can just use <F1> for a single and <F2>
for the triple, which is quite useful in markdown files.