Update pdfmerge to be a command instead of function
A reason why I hate shell scripting sometimes:

I was trying to use pdfmerge in xargs and was having difficulty with
filenames with spaces. I was properly using '-0'. The code was running
and attempting to merge but kept failing.

Turns out I had the old version of the command in the PATH still and so
xargs was using the command instead of the function. In the old version,
the parameters were passed as an unqouted $@, which is why the spaces in
the filenames ruined everything, even with the -0 option.

I didn't want to deal with exporting the function, so I just made it a
command again.