Add script to combine, rotate, and bookmark PDFs
We often get PDF sets for Mechanical and Architectural drawings that are
many files of single pages. They are also sometimes not rotated
properly. This script allows you to combine all the PDFs into one,
optionally apply a rotation to each of the pages, and then make a
bookmark for each page that is the original file name.