Add directory to begin storing awk functions
There are many cases in which there are common programming algorithms
that I will need to do, but awk doesn't have a built in function for.
The first example of this is rounding numbers. This new directory will
store these various functions that I can copy and paste into new
scripts.