Add notes and new macros for caption replacements
This whole thing is asinine. Word doesn't let you directly replace with
fields, so I'm using replacements using the clipboard. Went on an
internet searching journey looking for documentation on the wildcards
that are available in the find/replace. Ended up finding something in
the wayback machine for actual MS documentation that was worth
something.