Add dates.awk, update round.awk
I'm going to start taking advantage of the gawk specific feature to be
able to '@include' other awk files. In conjunction with the AWKPATH
variable, this is extremely powerful to be able to build up a library of
useful awk functions over time. For example, the date functions that are
in the dates.awk file in this commit.