Add SUDO_EDITOR environment variable
After some basic testing, it appears that for `sudoedit` to find an
editor to use, the EDITOR/VISUAL/SUDO_EDITOR variables must be
*absolute* paths, not just the command that can be looked up in the
path.

I don't feel like making the EDITOR and VISUAL values absolute paths, so
just left only the SUDO_EDITOR as an absolute, using `which` to
determine the path.