Remove 'encoding' from statusline
I can't believe I haven't noticed this until now. The 'encoding' option
is always 'utf-8' in Neovim. Always. No need to put it in the statusline.

There is the 'fileencoding' option which I might bring back in, but
honestly UTF-8 has taken over the world and it would mostly be just
noise.